FTYPE defaults to float (WAY faster), fpu fix

+++ common.h
@@ -1,6 +1,6 @@
 #ifndef COMMON_H
 #define COMMON_H
 
-typedef double FTYPE;
+typedef float FTYPE;
 
 #endif
